2nd

+++ src/Register.jsx
@@ -1,9 +1,9 @@
 import { GoogleLogin } from "@react-oauth/google";
-import { useEffect, useState } from "react";
+import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { loginUser } from "./firebaseAuth";
+import { registerUser } from "./firebaseAuth";
 
-export default function Login({ setUser }) {
+export default function Register({ setUser }) {
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -16,16 +16,21 @@ export default function Login({ setUser }) {
   const [message, setMessage] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
 
-  const handleLogin = async (e) => {
+  const handleRegister = async (e) => {
+    e.preventDefault();
+    if (password !== confirmPassword) {
+      setMessage("Passwords do not match.");
+      return;
+    }
     try {
-      e.preventDefault();
-      const token = await loginUser(email, password);
+      const token = await registerUser(email, password);
       localStorage.setItem("token", token);
       setUser(email);
       navigate("/");
     } catch (error) {
-      setMessage("Invalid email or password.");
+      setMessage("Invalid email or password");
     }
   };
 
@@ -33,9 +38,9 @@ export default function Login({ setUser }) {
     <div className="w-full bg-gray-950 flex justify-center items-center h-screen">
       <div className="w-96 mx-auto p-8 bg-white rounded-xl shadow-lg">
         <h1 className="text-black text-center mb-6 text-4xl font-bold">
-          Login
+          Register
         </h1>
-        <form onSubmit={handleLogin}>
+        <form onSubmit={handleRegister}>
           <label className="block text-gray-700 mb-2">Email:</label>
           <input
             type="email"
@@ -52,8 +57,16 @@ export default function Login({ setUser }) {
             value={password}
             onChange={(e) => setPassword(e.target.value)}
           />
-          <button className="w-full p-2 bg-blue-500 text-white rounded-lg hover:cursor-pointer hover:bg-blue-600 transition-colors duration-200 mb-4">
-            Login
+          <label className="block text-gray-700 mb-2">Confirm Password:</label>
+          <input
+            type="password"
+            placeholder="Confirm Password"
+            className="w-full p-2 mb-4 border border-gray-300 rounded-lg"
+            value={confirmPassword}
+            onChange={(e) => setConfirmPassword(e.target.value)}
+          />
+          <button className="w-full p-2 bg-blue-500 text-white rounded-lg hover:cursor-pointer hover:bg-blue-600 transition duration-200 mb-4">
+            Register
           </button>
           {message && (
             <p className="text-red-500 text-center text-sm mb-2">{message}</p>
@@ -62,16 +75,16 @@ export default function Login({ setUser }) {
         <h3 className="text-center text-gray-700 mb-4">OR</h3>
         <div className="flex justify-center">
           <GoogleLogin clientId="202469683498-0hf0bgijnv67do9hvj7idh6aijkdgg9b.apps.googleusercontent.com">
-            Login with Google
+            Register with Google
           </GoogleLogin>
         </div>
         <div className="flex justify-center mt-3">
-          <div className="flex text-center gap-1 text-sm">
-            New user? Click here to{" "}
-            <Link to="/register">
-              <p className="text-blue-500 text-sm">Register</p>
+          <p className="flex text-center gap-1 text-sm">
+            Have an account? Click here to
+            <Link to="/login">
+              <p className="text-blue-500 text-sm">Login</p>
             </Link>
-          </div>
+          </p>
         </div>
       </div>
     </div>
